Add validator methods

+++ server/src/controllers/listing.controller.ts
@@ -3,42 +3,51 @@ import { IListing } from '../models/listing/listing.model';
 import { IPet } from '../models/listing/pet/pet.model';
 import { IDog } from '../models/listing/pet/animal/dog/dog.model';
 
-const Listing = require('../models/listing/listing.model');
-const Pet = require('../models/listing/pet/pet.model');
-const Dog = require('../models/listing/pet/animal/dog/dog.model');
+const { Listing, itemTypes, saleTypes, mediaTypes } = require('../models/listing/listing.model');
+const { Pet, petTypes, genders } = require('../models/listing/pet/pet.model');
+const { Dog, sizes, hairCoats } = require('../models/listing/pet/animal/dog/dog.model');
 
 const createListing = async (req: Request, res: Response) => {
-  const { listerId, price, description, itemType, saleType, media } = req.body;
+  try {
+    const { listerId, price, description, itemType, saleType, media } = req.body;
 
-  await createItem(req, res)
-    .then(async (response: any) => {
-      if (!response) {
-        return ;
-      }
-      
-      const createdItem = response.item;
+    if (validateListing(req, res)) {
+      return ;
+    }
 
-      await Listing.create({
-        listerId,
-        itemId: createdItem[itemType.toLowerCase()]._id,
-        price,
-        description,
-        itemType,
-        saleType,
-        media
-      })
-        .then((listing: IListing) => {
-          console.log('Listing created successfully:', listing._id.toString());
-          return res.json({ listing: listing, item: createdItem });
+    await createItem(req, res)
+      .then(async (response: any) => {
+        if (!response) {
+          return ;
+        }
+        
+        const createdItem = response.item;
+
+        await Listing.create({
+          listerId,
+          itemId: createdItem[itemType.toLowerCase()]._id,
+          price,
+          description,
+          itemType,
+          saleType,
+          media
         })
-        .catch(async (error: Error) => {
-          req.params.id = createdItem[itemType.toLowerCase()]._id.toString();
-          req.params.itemType = itemType;
-          await deleteItem(req);
+          .then((listing: IListing) => {
+            console.log('Listing created successfully:', listing._id.toString());
+            return res.json({ listing: listing, item: createdItem });
+          })
+          .catch(async (error: any) => {
+            req.params.id = createdItem[itemType.toLowerCase()]._id.toString();
+            req.params.itemType = itemType;
+            await deleteItem(req);
 
-          return res.status(500).json({ message: error });
-        });
-    });
+            throw new Error(error.message);
+          });
+      });
+  } catch (error: any) {
+    console.log('Error creating Listing:\n' + error.message);
+    return res.status(500).json({ message: error.message });
+  }
 }
 
 const createItem = async (req: Request, res: Response) => {
@@ -50,37 +59,46 @@ const createItem = async (req: Request, res: Response) => {
 }
 
 const createPet = async (req: Request, res: Response) => {
-  const { animalType, originId, gender, birthdate } = req.body;
-  const { name } : { name?: string } = req.body;
+  try {
+    const { animalType, originId, gender, birthdate } = req.body;
+    const { name } : { name?: string } = req.body;
 
-  return await createAnimal(req, res)
-    .then(async (response: any) => {
-      if (!response) {
-        return ;
-      }
-      const createdAnimal = response.animal;
+    if (await validatePet(req, res)) {
+      return ;
+    }
 
-      return await Pet.create({
-        animalId: createdAnimal._id,
-        animalType,
-        originId,
-        name,
-        gender,
-        birthdate,
-      })
-        .then((pet: IPet) => {
-          console.log('Pet created successfully:', pet._id.toString());
-          return { item: { pet: pet, animal: createdAnimal }};
+    return await createAnimal(req, res)
+      .then(async (response: any) => {
+        if (!response) {
+          return ;
+        }
+
+        const createdAnimal = response.animal;
+
+        return await Pet.create({
+          animalId: createdAnimal._id,
+          animalType,
+          originId,
+          name,
+          gender,
+          birthdate,
         })
-        .catch(async (error: Error) => {
-          req.params.id = createdAnimal._id.toString();
-          req.params.animalType = animalType;
-          await deleteAnimal(req);
+          .then((pet: IPet) => {
+            console.log('Pet created successfully:', pet._id.toString());
+            return { item: { pet: pet, animal: createdAnimal }};
+          })
+          .catch(async (error: any) => {
+            req.params.id = createdAnimal._id.toString();
+            req.params.animalType = animalType;
+            await deleteAnimal(req);
 
-          let message = 'Error creating Pet: ' + error.message;
-          res.status(500).json({ message: message });
-        });
-    });
+            throw new Error(error.message);
+          });
+      });
+  } catch (error: any) {
+    console.log('Error creating Pet:');
+    throw new Error(error.message);
+  }
 }
 
 const createAnimal = async (req: Request, res: Response) => {
@@ -92,28 +110,32 @@ const createAnimal = async (req: Request, res: Response) => {
 }
 
 const createDog = async (req: Request, res: Response) => {
-  const { breedIds, weight, isMixedBreed, isMicrochipped, isNeutered, isPottyTrained } = req.body;
+  try {
+    const { breedIds, size, weight, hairCoat, isHypoallergenic, isMicrochipped, isNeutered, isPottyTrained, isHdbApproved } = req.body;
 
-  if (isMixedBreed && breedIds.length <= 1) {
-    return res.status(400).json({ message: 'Mixed breed dogs must have at least 2 breeds.' });
-  }
+    if (validateDog(req, res)) {
+      return ;
+    }
 
-  return await Dog.create({
-    breedIds,
-    weight,
-    isMixedBreed,
-    isMicrochipped,
-    isNeutered,
-    isPottyTrained
-  })
-    .then((dog: IDog) => {
-      console.log('Dog created successfully:', dog._id.toString());
-      return { animal: dog };
+    return await Dog.create({
+      breedIds,
+      size,
+      weight,
+      hairCoat,
+      isHypoallergenic,
+      isMicrochipped,
+      isNeutered,
+      isPottyTrained,
+      isHdbApproved
     })
-    .catch((error: Error) => {
-      let message = 'Error creating Dog: ' + error.message;
-      res.status(500).json({ message: message });
-    });
+      .then((dog: IDog) => {
+        console.log('Dog created successfully:', dog._id.toString());
+        return { animal: dog };
+      })
+  } catch (error: any) {
+    console.log('Error creating Dog:');
+    throw new Error(error.message);
+  }
 }
 
 const getAllListings = async (req: Request, res: Response) => {
@@ -122,9 +144,9 @@ const getAllListings = async (req: Request, res: Response) => {
       console.log('Listings retrieved successfully: [\n' + listings.map((listing: IListing) => listing._id.toString()).join(',\n') + '\n]');
       return res.json({ listings: listings });
     })
-    .catch((error: Error) => {
-      let message = 'Error retrieving Listings: ' + error.message;
-      return res.status(500).json({ message: message });
+    .catch((error: any) => {
+      console.log('Error retrieving Listings: ' + error)
+      return res.status(500).json({ message: error.message });
     });
 }
 
@@ -140,9 +162,9 @@ const getListing = async (req: Request, res: Response) => {
 
       return res.json({ listing: listing, item: await getItem(req, res) });
     })
-    .catch((error: Error) => {
-      let message = 'Error retrieving Listing: ' + error.message;
-      return res.status(500).json({ message: message });
+    .catch((error: any) => {
+      console.log('Error retrieving Listing:\n' + error);
+      return res.status(500).json({ message: error.message });
     });
 }
 
@@ -166,9 +188,9 @@ const getPet = async (req: Request, res: Response) => {
 
       return { pet: pet, animal: await getAnimal(req, res) };
     })
-    .catch((error: Error) => {
-      let message = 'Error retrieving Pet: ' + error.message;
-      return res.status(500).json({ message: message });
+    .catch((error: any) => {
+      console.log('Error retrieving Pet:\n' + error);
+      res.status(500).json({ message: error.message });
     });
 }
 
@@ -188,33 +210,34 @@ const getDog = async (req: Request, res: Response) => {
       console.log('Dog retrieved successfully:', dog._id.toString());
       return dog;
     })
-    .catch((error: Error) => {
-      let message = 'Error retrieving Dog:\n' + error.message;
-      return res.status(500).json({ message: message });
+    .catch((error: any) => {
+      console.log('Error retrieving Dog:\n' + error);
+      res.status(500).json({ message: error.message });
     });
 }
 
 const updateListing = async (req: Request, res: Response) => {
-  const { id } = req.params;
-  const { price, description, media } = req.body;
-
-  await Listing.findByIdAndUpdate(id, {
-    price,
-    description,
-    media
-  }, { new: true })
-    .then(async (updatedListing: IListing) => {
-      console.log('Listing updated successfully:', updatedListing._id.toString());
-      
-      req.params.id = updatedListing.itemId.toString();
-      req.params.itemType = updatedListing.itemType;
-      
-      return res.json({ listing: updatedListing, item: await updateItem(req) });
-    })
-    .catch((error: Error) => {
-      let message = 'Error updating Listing:\n' + error.message;
-      return res.json({ message: message });
-    });
+  try {
+    const { id } = req.params;
+    const { price, description, media } = req.body;
+  
+    await Listing.findByIdAndUpdate(id, {
+      price,
+      description,
+      media
+    }, { new: true })
+      .then(async (updatedListing: IListing) => {
+        console.log('Listing updated successfully:', updatedListing._id.toString());
+        
+        req.params.id = updatedListing.itemId.toString();
+        req.params.itemType = updatedListing.itemType;
+        
+        return res.json({ listing: updatedListing, item: await updateItem(req) });
+      })
+  } catch(error: any) {
+    console.log('Error updating Listing:\n' + error.message);
+    return res.json({ message: error.message });
+  };
 }
 
 const updateItem = async (req: Request) => {
@@ -226,25 +249,26 @@ const updateItem = async (req: Request) => {
 }
 
 const updatePet = async (req: Request) => {
-  const { id } = req.params;
-  const { name, birthdate } = req.body;
-
-  return await Pet.findByIdAndUpdate(id, {
-    name,
-    birthdate
-  }, { new: true })
-    .then(async (updatedPet: IPet) => {
-      console.log('Pet updated successfully:', updatedPet._id.toString());
-
-      req.params.id = updatedPet.animalId.toString();
-      req.params.animalType = updatedPet.animalType;
-
-      return { item: { pet: updatedPet, animal: await updateAnimal(req) }};
-    })
-    .catch((error: Error) => {
-      let message = 'Error updating Pet:\n' + error.message;
-      throw new Error(message);
-    });
+  try {
+    const { id } = req.params;
+    const { name, birthdate } = req.body;
+  
+    return await Pet.findByIdAndUpdate(id, {
+      name,
+      birthdate
+    }, { new: true })
+      .then(async (updatedPet: IPet) => {
+        console.log('Pet updated successfully:', updatedPet._id.toString());
+  
+        req.params.id = updatedPet.animalId.toString();
+        req.params.animalType = updatedPet.animalType;
+  
+        return { item: { pet: updatedPet, animal: await updateAnimal(req) }};
+      })
+  } catch(error: any) {
+    console.log('Error updating Pet:');
+    throw new Error(error.message);
+  };
 }
 
 const updateAnimal = async (req: Request) => {
@@ -256,23 +280,36 @@ const updateAnimal = async (req: Request) => {
 }
 
 const updateDog = async (req: Request) => {
-  const { id } = req.params;
-  const { weight, isMicrochipped, isNeutered, isPottyTrained } = req.body;
+  try {
+    const { id } = req.params;
+    const { size, weight, hairCoat, isHypoallergenic, isMicrochipped, isNeutered, isPottyTrained, isHdbApproved } = req.body;
 
-  return await Dog.findByIdAndUpdate(id, {
-    weight,
-    isMicrochipped,
-    isNeutered,
-    isPottyTrained
-  }, { new: true })
-    .then((updatedDog: IDog) => {
-      console.log('Dog updated successfully:', updatedDog._id.toString());
-      return { animal: updatedDog };
-    })
-    .catch((error: Error) => {
-      let message = 'Error updating Dog:\n' + error.message;
-      throw new Error(message);
-    });
+    if (size && !sizes.includes(size)) {
+      throw new Error('Invalid size: `' + size + '`');
+    }
+  
+    if (hairCoat && !hairCoats.includes(hairCoat)) {
+      throw new Error('Invalid hairCoat: `' + hairCoat + '`');
+    }
+  
+    return await Dog.findByIdAndUpdate(id, {
+      size,
+      weight,
+      hairCoat,
+      isHypoallergenic,
+      isMicrochipped,
+      isNeutered,
+      isPottyTrained,
+      isHdbApproved
+    }, { new: true })
+      .then((updatedDog: IDog) => {
+        console.log('Dog updated successfully:', updatedDog._id.toString());
+        return { animal: updatedDog };
+      })
+  } catch (error: any) {
+    console.log('Error updating Dog:');
+    throw new Error(error.message);
+  }
 }
 
 const deleteListing = async (req: Request, res: Response) => {
@@ -287,9 +324,9 @@ const deleteListing = async (req: Request, res: Response) => {
 
       return res.json({ listing: deletedListing, item: await deleteItem(req) });
     })
-    .catch((error: Error) => {
-      let message = 'Error deleting Listing:\n' + error.message;
-      return res.json({ message: message });
+    .catch((error: any) => {
+      console.log('Error deleting Listing:\n' + error.message)
+      return res.json({ message: error.message });
     });
 }
 
@@ -313,9 +350,9 @@ const deletePet = async (req: Request) => {
 
       return { item: { pet: deletedPet, animal: await deleteAnimal(req) }};
     })
-    .catch((error: Error) => {
-      let message = 'Error deleting Pet:\n' + error.message;
-      throw new Error(message);
+    .catch((error: any) => {
+      console.log('Error deleting Pet:\n' + error.message)
+      throw new Error(error.message);
     });
 }
 
@@ -335,10 +372,145 @@ const deleteDog = async (req: Request) => {
       console.log('Dog deleted successfully:', deletedDog._id.toString());
       return { animal: deletedDog };
     })
-    .catch((error: Error) => {
-      let message = 'Error deleting Dog:\n' + error.message;
-      throw new Error(message);
+    .catch((error: any) => {
+      console.log('Error deleting Dog:\n' + error.message)
+      throw new Error(error.message);
     });
+}
+
+const validateListing = (req: Request, res: Response) => {
+  const { listerId, price, description, itemType, saleType, media } = req.body;
+
+  if (!listerId) {
+    return res.status(400).json({ message: 'Missing listerId' });
+  }
+
+  if (!price) {
+    return res.status(400).json({ message: 'Missing price' });
+  }
+
+  if (isNaN(price)) {
+    return res.status(400).json({ message: 'Invalid price: `' + price + '`' });
+  }
+
+  if (!description) {
+    return res.status(400).json({ message: 'Missing description' });
+  }
+
+  if (!itemType) {
+    return res.status(400).json({ message: 'Missing itemType' });
+  }
+
+  if (!saleType) {
+    return res.status(400).json({ message: 'Missing saleType' });
+  }
+
+  if (!itemTypes.includes(itemType)) {
+    return res.status(400).json({ message: 'Invalid itemType: `' + itemType + '`' });
+  }
+
+  if (!saleTypes.includes(saleType)) {
+    return res.status(400).json({ message: 'Invalid saleType: `' + saleType + '`' });
+  }
+
+  if (media && media.length > 0) {
+    media.forEach((media: { type: string, url: string }) => {
+      if (!mediaTypes.includes(media.type)) {
+        return res.status(400).json({ message: 'Invalid media type: `' + media.type + '`' });
+      }
+    });
+  }
+
+  return false;
+}
+
+const validatePet = async (req: Request, res: Response) => {
+  const { animalType, originId, gender, birthdate } = req.body;
+
+  if (!animalType) {
+    return res.status(400).json({ message: 'Missing animalType' });
+  }
+
+  if (!originId) {
+    return res.status(400).json({ message: 'Missing originId' });
+  }
+  
+  if (!petTypes.includes(animalType)) {
+    return res.status(400).json({ message: 'Invalid animalType: `' + animalType + '`' });
+  }
+
+  if (!gender) {
+    return res.status(400).json({ message: 'Missing gender' });
+  }
+
+  if (!genders.includes(gender)) {
+    return res.status(400).json({ message: 'Invalid gender: `' + gender + '`' });
+  }
+
+  if (!birthdate) {
+    return res.status(400).json({ message: 'Missing birthdate' });
+  }
+
+  if (isNaN(Date.parse(birthdate))) {
+    return res.status(400).json({ message: 'Invalid birthdate: `' + birthdate + '`' });
+  }
+
+  return false;
+}
+
+const validateDog = (req: Request, res: Response) => {
+  const { size, weight, hairCoat, isHypoallergenic, isMicrochipped, isNeutered, isPottyTrained, isHdbApproved } = req.body;
+  const { breedIds } : { breedIds: string[] } = req.body;
+
+  if (!breedIds) {
+    return res.status(400).json({ message: 'Missing breedIds' });
+  }
+
+  if (!size) {
+    return res.status(400).json({ message: 'Missing size' });
+  }
+
+  if (!sizes.includes(size)) {
+    return res.status(400).json({ message: 'Invalid size: `' + size + '`' });
+  }
+
+  if (!weight) {
+    return res.status(400).json({ message: 'Missing weight' });
+  }
+
+  if (isNaN(weight)) {
+    return res.status(400).json({ message: 'Invalid weight: `' + weight + '`' });
+  }
+
+  if (!hairCoat) {
+    return res.status(400).json({ message: 'Missing hairCoat' });
+  }
+
+  if (!hairCoats.includes(hairCoat)) {
+    return res.status(400).json({ message: 'Invalid hairCoat: `' + hairCoat + '`' });
+  }
+
+  if (typeof isHypoallergenic === 'undefined') {
+    return res.status(400).json({ message: 'Missing isHypoallergenic' });
+  }
+
+  if (typeof isMicrochipped === 'undefined') {
+    return res.status(400).json({ message: 'Missing isMicrochipped' });
+  }
+
+  if (typeof isNeutered === 'undefined') {
+    return res.status(400).json({ message: 'Missing isNeutered' });
+  }
+
+  if (typeof isPottyTrained === 'undefined') {
+    return res.status(400).json({ message: 'Missing isPottyTrained' });
+  }
+
+  if (typeof isHdbApproved === 'undefined') {
+    return res.status(400).json({ message: 'Missing isHdbApproved' });
+  }
+
+  return false;
 }
 
 module.exports = {
